+++ wrappers/golang/pgadapter_test.go
@@ -30,7 +30,7 @@ func TestStart(t *testing.T) {
 		t.Fatalf("failed to get cache dir: %v", err)
 	}
 	downloadLocation := filepath.Join(home, "pgadapter-downloads")
-	for _, config := range []Config{
+	for i, config := range []Config{
 		{ExecutionEnvironment: &Docker{}},
 		{ExecutionEnvironment: &Docker{AlwaysPullImage: true}},
 		{ExecutionEnvironment: &Docker{}, Version: "v0.20.0"},
@@ -41,24 +41,24 @@ func TestStart(t *testing.T) {
 	} {
 		pgadapter, err := Start(context.Background(), config)
 		if err != nil {
-			t.Fatalf("failed to start PGAdapter: %v", err)
+			t.Fatalf("%d: failed to start PGAdapter: %v", i, err)
 		}
 		port, err := pgadapter.GetHostPort()
 		if err != nil {
-			t.Fatalf("failed to get port from PGAdapter: %v", err)
+			t.Fatalf("%d: failed to get port from PGAdapter: %v", i, err)
 		}
 		if port == 0 {
-			t.Fatal("Port should be non-zero", err)
+			t.Fatalf("%d: Port should be non-zero: %v", i, err)
 		}
 
 		servAddr := fmt.Sprintf("localhost:%d", port)
 		tcpAddr, err := net.ResolveTCPAddr("tcp", servAddr)
 		if err != nil {
-			t.Fatalf("ResolveTCPAddr failed: %v", err)
+			t.Fatalf("%d: ResolveTCPAddr failed: %v", i, err)
 		}
 		conn, err := net.DialTCP("tcp", nil, tcpAddr)
 		if err != nil {
-			t.Fatalf("Dial failed: %v", err)
+			t.Fatalf("%d: Dial failed: %v", i, err)
 		}
 		// Send an invalid startup message. This should validate that we are talking to
 		// a working PGAdapter instance.